+++ src/app/data_harvester/disks/unix.rs
@@ -18,13 +18,14 @@ cfg_if::cfg_if! {
     }
 }
 
+use super::{keep_disk_entry, DiskHarvest, IoData, IoHarvest};
 use crate::app::Filter;
-use crate::data_harvester::disks::{DiskHarvest, IoData, IoHarvest};
 
+/// Returns the I/O usage of certain mount points.
 pub fn get_io_usage() -> anyhow::Result<IoHarvest> {
     let mut io_hash: HashMap<String, Option<IoData>> = HashMap::new();
 
-    for io in io_stats()?.flatten() {
+    for io in io_stats()?.into_iter().flatten() {
         let mount_point = io.device_name().to_string_lossy();
 
         io_hash.insert(
@@ -39,6 +40,7 @@ pub fn get_io_usage() -> anyhow::Result<IoHarvest> {
     Ok(io_hash)
 }
 
+/// Returns the disk usage of the mounted (and for now, physical) disks.
 pub fn get_disk_usage(
     disk_filter: &Option<Filter>, mount_filter: &Option<Filter>,
 ) -> anyhow::Result<Vec<DiskHarvest>> {
@@ -55,43 +57,7 @@ pub fn get_disk_usage(
         // 2. Is the entry denied through any filter? That is, does it match an entry in a filter where `is_list_ignored` is `true`? If so, we always deny this entry.
         // 3. Anything else is allowed.
 
-        let filter_check_map = [(disk_filter, &name), (mount_filter, &mount_point)];
-
-        // This represents case 1. That is, if there is a match in an allowing list - if there is, then
-        // immediately allow it!
-        let matches_allow_list = filter_check_map.iter().any(|(filter, text)| {
-            if let Some(filter) = filter {
-                if !filter.is_list_ignored {
-                    for r in &filter.list {
-                        if r.is_match(text) {
-                            return true;
-                        }
-                    }
-                }
-            }
-            false
-        });
-
-        let to_keep = if matches_allow_list {
-            true
-        } else {
-            // If it doesn't match an allow list, then check if it is denied.
-            // That is, if it matches in a reject filter, then reject.  Otherwise, we always keep it.
-            !filter_check_map.iter().any(|(filter, text)| {
-                if let Some(filter) = filter {
-                    if filter.is_list_ignored {
-                        for r in &filter.list {
-                            if r.is_match(text) {
-                                return true;
-                            }
-                        }
-                    }
-                }
-                false
-            })
-        };
-
-        if to_keep {
+        if keep_disk_entry(&name, &mount_point, disk_filter, mount_filter) {
             // The usage line can fail in some cases (for example, if you use Void Linux + LUKS,
             // see https://github.com/ClementTsang/bottom/issues/419 for details).
             if let Ok(usage) = partition.usage() {